Fix role-probe pipeline to match reference: layer, skip_first_n, dtype
- Probe post_attention_layernorm's output (pre-MLP hidden state), not the
  decoder block's full output -- the reference trains on
  all_pre_mlp_hidden_states, a different tensor entirely.
- Add the skip_first_n=32 turn-position filter the reference applies for
  nested_reasoning, dropping each labeled turn's first 32 tokens.
- Normalize target_role "thinking" -> "cot" so chain-of-thought samples
  actually contribute labeled tokens instead of being silently dropped.
- Load the model with dtype="auto" instead of the implicit float32 default,
  and round captured activations to float16 at capture time, matching the
  reference's precision.
- Guard ProbeTrainer._fit against a role missing from the train split, and
  add debug logging (role/token counts, fit shapes) to compare runs.
- Fix InMemoryTokenActivationDataset.summary()'s activation norm overflowing
  in float16.

+++ src/nnact/_outputs/_dataset.py
@@ -250,7 +250,13 @@ class InMemoryTokenActivationDataset(ActivationDataset):
         columns["predicted_id"] = self.prediction.tolist()
 
         for name, tensor in self.activations.items():
-            columns[f"{name}_norm"] = np.linalg.norm(tensor, axis=-1).tolist()
+            # float16 activations can exceed ~65504 in a transformer's
+            # residual stream; squaring them in-dtype for the norm overflows
+            # to inf, so compute the norm in float32 regardless of storage
+            # dtype.
+            columns[f"{name}_norm"] = np.linalg.norm(
+                tensor.astype(np.float32), axis=-1
+            ).tolist()
 
         return pd.DataFrame(
             columns, index=pd.RangeIndex(int(offsets[-1]), name="token")
